moved navbar components to components folder
